Move IntoVhdl trait into til_vhdl crate root
